Watch and settings area
I added a settings layout, and I've made a watch layout. The watch
layout is clickable, and I was thinking we could allow the user to start
and stop the tracking from their watch (double tap).
Anyways, I'm not sure how to force the watch interface to work..

Also, I'm going to try to work on this later. Bed is calling me :P